+++ src/Entity/TokenInterface.php
@@ -3,18 +3,19 @@
 namespace Ifrost\DoctrineApiAuthBundle\Entity;
 
 use Ifrost\DoctrineApiBundle\Entity\EntityInterface;
+use Ramsey\Uuid\UuidInterface;
 
 interface TokenInterface extends EntityInterface
 {
-    public function getUuid(): string;
+    public function getUuid(): UuidInterface;
 
-    public function getUserUuid(): string;
+    public function getUserUuid(): UuidInterface;
+
+    public function getRefreshTokenUuid(): UuidInterface;
 
     public function getIat(): int;
 
     public function getExp(): int;
 
     public function getDevice(): string;
-
-    public function getRefreshTokenUuid(): string;
 }