+++ tests/Unit/BundleTestCase.php
@@ -15,6 +15,7 @@ use Ifrost\DoctrineApiBundle\Query\Entity\EntityQuery;
 use Ifrost\DoctrineApiBundle\Utility\DbClient;
 use PHPUnit\Framework\TestCase;
 use PlainDataTransformer\Transform;
+use Ramsey\Uuid\Uuid;
 
 class BundleTestCase extends TestCase
 {
@@ -30,18 +31,18 @@ class BundleTestCase extends TestCase
         ]);
         $this->db = new DbClient($this->dbal);
         $this->user = User::createFromArray([
-            'uuid' => '3fc713ae-f1b8-43a6-95d2-e6d573fab41a',
+            'uuid' => Uuid::fromString('3fc713ae-f1b8-43a6-95d2-e6d573fab41a'),
             'email' => 'tom.smith@email.com',
             'roles' => ['ROLE_USER'],
             'password' => '123'
         ]);
         $this->token = Token::createFromArray([
-            'uuid' => '25feb06b-0c1e-4416-86fc-706134f2c7de',
-            'user_uuid' => '3fc713ae-f1b8-43a6-95d2-e6d573fab41a',
+            'uuid' => Uuid::fromString('25feb06b-0c1e-4416-86fc-706134f2c7de'),
+            'user_uuid' => Uuid::fromString('3fc713ae-f1b8-43a6-95d2-e6d573fab41a'),
+            'refresh_token_uuid' => Uuid::fromString('60efd5f1-d831-4c02-863d-4ee11843fc2e'),
             'iat' => 1573457094,
             'exp' => 1573460694,
             'device' => '',
-            'refresh_token_uuid' => '60efd5f1-d831-4c02-863d-4ee11843fc2e',
         ]);
     }
 
@@ -53,14 +54,11 @@ class BundleTestCase extends TestCase
     protected function createUserIfNotExists(User $user): void
     {
         try {
-            $this->db->fetchColumn(EntityQuery::class, $user::getTableName(), $user->getUuid());
+            $this->db->fetchColumn(EntityQuery::class, $user::getTableName(), $user->getUuid()->getBytes());
         } catch (Exception $e) {
             throw $e;
         } catch (NotFoundException) {
-            $data = $user->jsonSerialize();
-            $data['roles'] = json_encode($data['roles']);
-            $data['password'] = $user->getPassword();
-            $this->db->insert($user::getTableName(), $data);
+            $this->db->insert($user::getTableName(), $user->getWritableFormat());
         }
     }
 }